requirements.txt
fastapi==0.115.12
mangum==0.19.0
boto3==1.38.12
itsdangerous==2.2.0
python-multipart==0.0.20
aws-cdk-lib==2.189.1
constructs>=10.0.0,<11.0.0
